refactor(ui): TakvimAyarlari sunum bilesenlerini ayir (Faz E)
OzetSatir, takvimUygulamasiniAc, BildirimBanneri, OzellikBilgi, BasariIcerigi
-> screens/TakvimAyarlari/bilesenler.tsx. TakvimAyarlari 1727 -> 1449 satir.
Orphan Linking/Platform importlari temizlendi.

391 test geciyor, 0 tip hatasi.
